KAFKA-10243; ConcurrentModificationException while processing connection setup timeouts (#8990)
This PR fixes a bug introduced in #8683.

While processing connection set up timeouts, we are iterating through the connecting nodes to process timeouts and we disconnect within the loop, removing the entry from the set in the loop that it iterating over the set. That raises a ConcurrentModificationException exception. The current unit test did not catch this because it was using only one node.

Reviewers: Rajini Sivaram <rajinisivaram@googlemail.com>